devfolder/.system/devfolder_backup/backup_20240717-021315/app/services/ai_model_service.py

# app/services/ai_model_service.py

"""
Handles interactions with AI models.

This module provides a service for loading, generating responses from,
and fine-tuning AI models using the Anthropic API.
"""

from typing import Any, List, Dict
from app.utils.api_utils import AsyncAnthropic
import os
import logging
import traceback

class AIModelService:
    def __init__(self):
        self.anthropic = AsyncAnthropic(api_key=os.environ.get("ANTHROPIC_API_KEY"))
        self.logger = logging.getLogger(__name__)

    async def load_model(self, model_name: str) -> Any:
        """
        Loads an AI model.

        Args:
            model_name (str): The name of the model to load.

        Returns:
            Any: The loaded model object.

        Raises:
            ValueError: If the model_name is not supported.
        """
        try:
            # For Anthropic models, we don't need to explicitly load them
            # We'll just return the model name as it will be used in generate_response
            supported_models = ["gpt-4", "claude-2", "claude-instant-1"]
            if model_name not in supported_models:
                raise ValueError(f"Unsupported model: {model_name}")
            return model_name
        except Exception as e:
            self.logger.error(f"Error loading model {model_name}: {str(e)}")
            self.logger.debug(traceback.format_exc())
            raise

    async def generate_response(self, prompt: str, model: Any) -> str:
        """
        Generates a response using the specified AI model.

        Args:
            prompt (str): The input prompt for the model.
            model (Any): The model object to use for generation.

        Returns:
            str: The generated response.

        Raises:
            Exception: If there's an error during response generation.
        """
        try:
            response = await self.anthropic.completions.create(
                model=model,
                prompt=prompt,
                max_tokens_to_sample=1000,
            )
            return response.completion
        except Exception as e:
            self.logger.error(f"Error generating response: {str(e)}")
            self.logger.debug(traceback.format_exc())
            raise

    async def fine_tune_model(self, model: Any, training_data: List[Dict[str, str]]) -> Any:
        """
        Fine-tunes the specified AI model using the provided training data.

        Args:
            model (Any): The model object to fine-tune.
            training_data (List[Dict[str, str]]): A list of training examples.

        Returns:
            Any: The fine-tuned model object.

        Raises:
            NotImplementedError: As fine-tuning is not currently supported by the Anthropic API.
        """
        try:
            # Currently, Anthropic doesn't support fine-tuning through their API
            # This method is included for future compatibility
            raise NotImplementedError("Fine-tuning is not currently supported by the Anthropic API")
        except Exception as e:
            self.logger.error(f"Error fine-tuning model: {str(e)}")
            self.logger.debug(traceback.format_exc())
            raise

    async def get_model_info(self, model_name: str) -> Dict[str, Any]:
        """
        Retrieves information about a specific model.

        Args:
            model_name (str): The name of the model.

        Returns:
            Dict[str, Any]: A dictionary containing model information.

        Raises:
            ValueError: If the model_name is not supported.
        """
        try:
            supported_models = {
                "gpt-4": {"type": "GPT", "version": "4", "capabilities": ["text generation", "code completion"]},
                "claude-2": {"type": "Claude", "version": "2", "capabilities": ["text generation", "analysis"]},
                "claude-instant-1": {"type": "Claude", "version": "Instant 1", "capabilities": ["fast responses", "text generation"]}
            }
            if model_name not in supported_models:
                raise ValueError(f"Unsupported model: {model_name}")
            return supported_models[model_name]
        except Exception as e:
            self.logger.error(f"Error retrieving model info for {model_name}: {str(e)}")
            self.logger.debug(traceback.format_exc())
            raise

if __name__ == "__main__":
    # This block is for testing purposes only
    import asyncio

    async def test_ai_model_service():
        service = AIModelService()
        model = await service.load_model("claude-2")
        response = await service.generate_response("Hello, how are you?", model)
        print(f"Generated response: {response}")

        model_info = await service.get_model_info("claude-2")
        print(f"Model info: {model_info}")

    asyncio.run(test_ai_model_service())
